moved player files into folder
